+++ src/main/java/com/sounganization/botanify/common/config/restTemplate/RestTemplateConfig.java
@@ -1,0 +1,18 @@
+package com.sounganization.botanify.common.config.restTemplate;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+import org.springframework.web.util.DefaultUriBuilderFactory;
+
+@Configuration
+public class RestTemplateConfig {
+    @Bean(name = "locationRestTemplate")
+    public RestTemplate locationRestTemplate() {
+        RestTemplate restTemplate = new RestTemplate();
+        DefaultUriBuilderFactory factory = new DefaultUriBuilderFactory();
+        factory.setEncodingMode(DefaultUriBuilderFactory.EncodingMode.URI_COMPONENT);
+        restTemplate.setUriTemplateHandler(factory);
+        return restTemplate;
+    }
+}
